Addition of ReaLife token (#439)
* Addition of ReaLife token

Token to be used within the ReaLife game, currently in development.

* Adding Tokens for DTF app

DTF...The ultimate dating application! Bringing people together through technology, now that is awesome. However, current options for dating apps leave a little (or a lot) to be desired. DTF takes things to a whole NEW level. From mild to wild, and everything in between. If you are looking for a fling, or searching for your soulmate, DTF has you covered!

* Adding Hydrogen Token

HYD will be used as the utility token for three phases of the LifeRhythm Platform.

* lowercase